Extract NumberInputBase with shared layout logic
